+++ src/xpcs/filter/dense_filter.cpp
@@ -109,7 +109,7 @@ DenseFilter::~DenseFilter() {
 
 void DenseFilter::Apply(struct xpcs::io::ImmBlock* blk) {
   int **indx = blk->index;
-  short **val = blk->value;
+  float **val = blk->value;
   int frames = blk->frames;
 
   double *dark_avg = NULL;
@@ -125,7 +125,7 @@ void DenseFilter::Apply(struct xpcs::io::ImmBlock* blk) {
   for (int i = 0; i < frames; i++) {
     int pixels = ppf[i];
     int *index = indx[i];
-    short *value = val[i];
+    float *value = val[i];
     float f_sum = 0.0f;
     int pix_cnt = frame_width_ * frame_height_;
     int sbin = 0;
